Se implemento la base para el control de acceso por modulos

+++ application/views/admin/categorias/list.php
@@ -13,9 +13,25 @@
                 <!-- Default box -->
                 <div class="box box-solid">
                     <div class="box-body">
+
+                        <div class="row text-center">
+                            <div class="col-md-4"><?php if ($permisos->insert == 0){ echo '<h4><span class="label label-danger ">No puede Insertar, permiso denegado!!</span></h4>';} ?></div>
+                            <div class="col-md-4"><?php if ($permisos->update == 0){ echo '<h4><span class="label label-danger">No puede Actualizar, permiso denegado!!</span></h4>';} ?>
+                            </div>
+                            <div class="col-md-4"><?php if ($permisos->delete == 0){ echo '<h4><span class="label label-danger">No puede Eliminar, permiso denegado!!</span></h4>';} ?>
+                            </div>
+                        </div>
+
                         <div class="row">
                           <div class="col-md-12 text-center">
+                            
+                            <!-- VALIDANDO PERMISOS -->
+                            
+                            <!-- FIN VALIDANDO PERMISOS -->
+
+                            <?php if ($permisos->insert == 1): ?>
                               <a href="<?php echo base_url();?>mantenimiento/Ccategorias/addCat" class="btn btn-success btn-flat"><span class="fa fa-plus"></span>Agregar Categorias</a>
+                            <?php endif; ?>
                           </div>  
                         </div>
                         <hr>
@@ -46,8 +62,14 @@
                                                     <button type="button" class="btn btn-info btn-view" data-toggle="modal" data-target="#modal-default" value="<?php echo $cat->id; ?>">
                                                     <span class="fa fa-eye"></span>
                                                     </button>
-                                                     <a href="<?php echo base_url(); ?>mantenimiento/Ccategorias/preUpdate/<?php echo $cat->id;?>" class="btn btn-warning"><span class="fa fa-pencil"></span></a>
-                                                     <a href="<?php echo base_url(); ?>mantenimiento/Ccategorias/delete/<?php  echo $cat->id;?>" class="btn btn-danger btn-delete" value="<?php echo $cat->id;?>"><span class="fa fa-remove"></span></a>
+
+                                                    <?php if ($permisos->update == 1): ?>
+                                                         <a href="<?php echo base_url(); ?>mantenimiento/Ccategorias/preUpdate/<?php echo $cat->id;?>" class="btn btn-warning"><span class="fa fa-pencil"></span></a>
+                                                    <?php endif; ?>
+
+                                                    <?php if ($permisos->delete == 1): ?>
+                                                         <a href="<?php echo base_url(); ?>mantenimiento/Ccategorias/delete/<?php  echo $cat->id;?>" class="btn btn-danger btn-delete" value="<?php echo $cat->id;?>"><span class="fa fa-remove"></span></a>
+                                                    <?php endif; ?>
                                                 </div>
                                             </td>
                                         </tr>
